Elasticsearch: Rename public/app/plugins/datasource/elasticsearch/query_builder.ts to follow convention
