+++ public/app/plugins/datasource/elasticsearch/QueryBuilder.ts
